roachtest: move clusterSpec to its own file
It will need some massaging (it references pkg singletons) before it
can be moved out into a subpackage.

Release note: None
